Move PaywallPresenter to its own file
