+++ Tests/TestingApps/PaywallsTester/PaywallsTester/UI/Views/PaywallPresenter.swift
@@ -1,0 +1,39 @@
+//
+//  PaywallPresenter.swift
+//  PaywallsTester
+//
+//  Created by James Borthwick on 2024-04-25.
+//
+
+import SwiftUI
+import RevenueCat
+#if DEBUG
+@testable import RevenueCatUI
+#else
+import RevenueCatUI
+#endif
+
+struct PaywallPresenter: View {
+
+    var offering: Offering
+    var mode: PaywallViewMode
+    var displayCloseButton: Bool = Configuration.defaultDisplayCloseButton
+
+    var body: some View {
+        switch self.mode {
+        case .fullScreen:
+            PaywallView(offering: self.offering, displayCloseButton: self.displayCloseButton)
+
+        #if !os(watchOS)
+        case .footer:
+            CustomPaywallContent()
+                .paywallFooter(offering: self.offering)
+
+        case .condensedFooter:
+            CustomPaywallContent()
+                .paywallFooter(offering: self.offering, condensed: true)
+        #endif
+        }
+    }
+
+}
